+++ app/Models/User.php
@@ -88,4 +88,8 @@ class User extends Authenticatable
     {
         return $this->name.' <'.$this->email.'>';
     }
+
+    public function getMember(): Member|null {
+        return Member::query()->where("email", $this->email)->first();
+    }
 }